moved normalizeConnectorId to designer utils since will be re-using in designer
