+++ libs/designer/src/lib/ui/templates/connections/displayConnections.tsx
@@ -15,7 +15,7 @@ import { Text } from '@fluentui/react-components';
 import type { Connection, Template } from '@microsoft/logic-apps-shared';
 import { ConnectionService, getObjectPropertyValue } from '@microsoft/logic-apps-shared';
 import type { AppDispatch, RootState } from '../../../core/state/templates/store';
-import { getConnectorResources, normalizeConnectorId } from '../../../core/templates/utils/helper';
+import { getConnectorResources } from '../../../core/templates/utils/helper';
 import { type IntlShape, useIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 import { ConnectorIconWithName } from './connector';
@@ -26,7 +26,7 @@ import React, { useEffect, useState } from 'react';
 import { updateTemplateConnection } from '../../../core/actions/bjsworkflow/connections';
 import { getConnector } from '../../../core/queries/operation';
 import type { ConnectorInfo } from '../../../core/templates/utils/queries';
-import { isConnectionValid } from '../../../core/utils/connectors/connections';
+import { isConnectionValid, normalizeConnectorId } from '../../../core/utils/connectors/connections';
 
 const createPlaceholderKey = '##create##';
 const connectionStatus: Record<string, any> = {
